pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relay-teams"
dynamic = ["version"]
description = "Role-driven multi-agent framework with strong typing and tool-based collaboration."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7.0",
  "pydantic-ai",
  "pydantic-ai-slim[fastmcp]",
  "PyYAML>=6.0.1",
  "typer>=0.12.3",
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.23.0",
  "httpx>=0.27.0",
  "json-repair>=0.58.7",
  "lark-oapi==1.5.3",
  "markdownify>=1.2.0",
  "defusedxml>=0.7.1",
  "qrcode>=8.0",
  "pywinpty>=3.0.3; platform_system == 'Windows'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-timeout >= 2.1.0",
  "pytest-asyncio >= 1.2.0",
  "coverage >= 7.8.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
  "playwright>=1.49.0",
  "pre-commit >= 4.5.0",
  "ruff>=0.9.10",
  "basedpyright>=1.36.2",
  "pyright>=1.1.408",
]

[project.scripts]
relay-teams = "relay_teams.interfaces.cli.app:main"
relay-teams-evals = "relay_teams_evals.run:app"

[tool.pytest.ini_options]
testpaths = ["tests/unit_tests", "tests/integration_tests"]
pythonpath = ["src", "tests", "."]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "relay_teams._version.__version__" }

[tool.setuptools.package-data]
"relay_teams.builtin" = [
  "config/*.json",
  "logging/*.ini",
  "roles/*.md",
  "skills/**/*",
]
"relay_teams.tools" = ["*/*.txt"]

[dependency-groups]
evals = [
  "chardet<6",
  "swebench>=4.0.0",
  "docker>=7.0.0",
  "datasets>=2.14.0",
]

[tool.uv]
default-groups = ["evals"]

[tool.pyright]
extraPaths = ["src", "tests"]
typeCheckingMode = "standard"
exclude = [
  ".venv",
  "**/__pycache__",
  "**/node_modules",
  "src/relay_teams/builtin/skills/pptx-craft/scripts/svg_to_pptx",
]

[tool.ruff]
extend-exclude = ["src/relay_teams/builtin/skills/pptx-craft/scripts/svg_to_pptx"]

